add renderer class

+++ console_game/Goal.cpp
@@ -6,7 +6,7 @@ Goal::Goal(short x, short y)
 	curr_pos = COORD{x, y};
 	prev_pos = COORD{x, y};
 	graphic = 'G';
-	foreground_color = static_cast<WORD>(Color::BLACK);
-	background_color = static_cast<WORD>(Color::LIGHT_SKY_BLUE);
+	foreground_color = Color::BLACK;
+	background_color = Color::LIGHT_SKY_BLUE;
 	hp = -1;
 }
